+++ src/tui/new_request.rs
@@ -1,5 +1,5 @@
 mod kvd_section;
-mod new_request;
+mod wizard;
 
 pub(crate) use kvd_section::*;
-pub(crate) use new_request::*;
+pub(crate) use wizard::*;